requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.5.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Chroma==0.2.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.18
commonmark==0.9.1
dataclasses-json==0.5.7
Deprecated==1.2.13
duckdb==0.7.1
et-xmlfile==1.1.0
fastapi==0.95.0
filelock==3.10.7
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
kor==0.6.1
langchain==0.0.131
llama-index==0.5.7
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
olefile==0.46
openai==0.27.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
posthog==2.4.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyDrive==1.3.1
Pygments==2.14.0
PyMySQL==1.0.3
pypandoc==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
rdflib==6.3.2
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.9
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zstandard==0.20.0
